Move the controller handler to the `controller.py` file
Remove `controller_handler.py`.
It makes it easier for testing.
