+++ server/api/crud/model_monitoring/deployment.py
@@ -41,7 +41,7 @@ _STREAM_PROCESSING_FUNCTION_PATH = (
 _MONITORING_ORIGINAL_BATCH_FUNCTION_PATH = _MODEL_MONITORING_COMMON_PATH / "batch.py"
 
 _MONITORING_APPLICATION_CONTROLLER_FUNCTION_PATH = (
-    _MODEL_MONITORING_COMMON_PATH / "controller_handler.py"
+    _MODEL_MONITORING_COMMON_PATH / "controller.py"
 )
 
 _MONITORING_WRITER_FUNCTION_PATH = _MODEL_MONITORING_COMMON_PATH / "writer.py"
